requirements.txt
apache-airflow-providers-snowflake
python-dotenv
requests
astronomer-cosmos

